requirements.txt
gradio
requests
langgraph
langchain-openai
dotenv
langchain-community
ddgs
langfuse
wikipedia